Fix failing test testConnectors()
There was a discrepancy between the amount of errors when checking all cocos and just the single coco. This was due to the fact that the connector symbol is logging errors in getPort(). The error thus occured more often with all cocos as multiple cocos tried to use the getPort() method on a non existing port.