Add update() call to initialization of atomic components
Generated file for atomic components will now call the update() method after setting the result. This should ensure that the values in output ports that were set in the initial transition are used in the first actual cycle. Otherwise the values would always not be read, as they were only the next value of the port and then overwritten in the first transition.